#34438 Split member enums into one-type-per-file
Move the 7 enums from ConMember.cs into individual files (matching the
folder convention for ConLoadEffectPositionEnum etc.). No type or
namespace changes; purely file organization.
