+++ src/IdeaStatiCa.Api/Connection/Model/Connection/ConAlignedPlateSideCodeEnum.cs
@@ -1,0 +1,11 @@
+namespace IdeaStatiCa.Api.Connection.Model
+{
+	public enum ConAlignedPlateSideCodeEnum : int
+	{
+		Default = 0,
+		LowerSide = 1,
+		UpperSide = 2,
+		BothSides = 3,
+		Center = 4
+	}
+}
