+++ src/IdeaStatiCa.Api/Connection/Model/Connection/ConMember.cs
@@ -138,90 +138,4 @@ namespace IdeaStatiCa.Api.Connection.Model
 		[JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
 		public double? TheoreticalLengthY { get; set; }
 	}
-
-	public enum ConAlignedPlateSideCodeEnum : int
-	{
-		Default = 0,
-		LowerSide = 1,
-		UpperSide = 2,
-		BothSides = 3,
-		Center = 4
-	}
-
-	public enum ConMemberForcesInEnum
-	{
-		Position = 0,
-		Node = 1,
-		Bolts = 2,
-		SelectedMemberFace = 3,
-	}
-
-	/// <summary>
-	/// Which end of the member is connected in the joint.
-	/// </summary>
-	public enum ConMemberConnectedByEnum
-	{
-		Begin = 0,
-		End = 1,
-	}
-
-	/// <summary>
-	/// Analysis model type of a member - which force/moment components are transferred.
-	/// Names and values mirror the internal BeamSegmentModelType so the two convert by name.
-	/// </summary>
-	public enum ConMemberModelTypeEnum
-	{
-		/// <summary>N-Vy-Vz-Mx-My-Mz</summary>
-		LoadedInXYZ = 0,
-		/// <summary>N-Vz-My</summary>
-		LoadedInXZ = 2,
-		/// <summary>N-Vy-Mz</summary>
-		LoadedInXY = 4,
-		/// <summary>N-Vy-Vz</summary>
-		LoadedInXYZ_NoBending = 32,
-	}
-
-	public enum ConMemberPlacementDefinitionTypeEnum
-	{
-		None = 0,
-		DirectionVector = 1,
-		RotationsOfX = 2,
-		Lcs,
-	}
-
-	public enum ConMemberAlignmentTypeEnum
-	{
-		None,
-		ToMemberPlate,
-		ToMemberPlateRotateThenTranslate,
-	}
-
-	public enum ConMemberPlatePartTypeEnum : int
-	{
-		NotSpecified = 0,
-		TopFlange = 1,
-		BottomFlange = 2,
-		Web = 4,
-		BasePlate = 8,
-		EndPlate = 16,
-		PlateWidener = 32,
-		Stiffener = 64,
-		Rib = 128,
-		GussetPlate = 256,
-		FinPlate = 512,
-		Flange = 1024,
-		CssArcSegment = 2048,
-		IsStub = 4096,
-		Splice = 8192,
-		TonguePlate = 16384,
-		LidPlate = 32768,
-		GeneralPlate = 65536,
-		Doubler = GeneralPlate * 2,
-		EndPlateOnFlanges = Doubler * 2,
-		BackingPlate = EndPlateOnFlanges * 2,
-		InsertedPlate = BackingPlate * 2,
-		IsNegative = InsertedPlate * 2,
-		BothFlanges = 3,
-		AllCssParts = 7
-	}
 }
