+++ src/IdeaStatiCa.Api/Connection/Model/Connection/ConMemberAlignmentTypeEnum.cs
@@ -1,0 +1,9 @@
+namespace IdeaStatiCa.Api.Connection.Model
+{
+	public enum ConMemberAlignmentTypeEnum
+	{
+		None,
+		ToMemberPlate,
+		ToMemberPlateRotateThenTranslate,
+	}
+}
