+++ src/IdeaStatiCa.Api/Connection/Model/Connection/ConMemberConnectedByEnum.cs
@@ -1,0 +1,11 @@
+namespace IdeaStatiCa.Api.Connection.Model
+{
+	/// <summary>
+	/// Which end of the member is connected in the joint.
+	/// </summary>
+	public enum ConMemberConnectedByEnum
+	{
+		Begin = 0,
+		End = 1,
+	}
+}
