+++ src/IdeaStatiCa.Api/Connection/Model/Connection/ConMemberForcesInEnum.cs
@@ -1,0 +1,10 @@
+namespace IdeaStatiCa.Api.Connection.Model
+{
+	public enum ConMemberForcesInEnum
+	{
+		Position = 0,
+		Node = 1,
+		Bolts = 2,
+		SelectedMemberFace = 3,
+	}
+}
