+++ src/IdeaStatiCa.Api/Connection/Model/Connection/ConMemberPlacementDefinitionTypeEnum.cs
@@ -1,0 +1,10 @@
+namespace IdeaStatiCa.Api.Connection.Model
+{
+	public enum ConMemberPlacementDefinitionTypeEnum
+	{
+		None = 0,
+		DirectionVector = 1,
+		RotationsOfX = 2,
+		Lcs,
+	}
+}
